pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    "python-frontmatter (==1.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "pyside6 (>=6.9.0,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
